Fix for window resizing and better layout of Device window
In this commit, we have removed the font setting, 40 points is
a REALLY big font and made all controls gigantic. Removing that
setting (you should really let the desktop theme decide) made
it work as expected.

Also, I took the liberty of creating a better layout for the controls
(feel free to return to the previous one) that allows to really put
and emphasis on the graph